+++ ports/stm32/boards/Passport/modules/tests/test_unit.py
@@ -50,3 +50,7 @@ def test_multisig_xpub_validation(test):
 
 def test_psbt_fee(test):
     assert test('psbt_fee.py') == b'OK'
+
+
+def test_unchained(test):
+    assert test('unchained.py') == b'OK'